Another couple of client interface fixes:
i) Ensure that the CIF caller's idea of pb->nret is always respected, but if there is a mismatch then log a warning if DEBUG_CIF 
is enabled, then correct the stack and continue.

ii) Correct the code path for call-method and interpret to ensure they ignore the extra status variable pushed onto the stack by 
client-call-iface.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@siriusit.co.uk>


git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@733 f158a5a8-5612-0410-a976-696ce0be7e32