ppc: Increase #address-cells for ppc64
To store a full 64-bit address we need two cells. This setting matches
Apple's OpenFirmware on my G5 and IBM's firmware on the JS20.

The PCI driver zero-pads its physical addresses, so is okay as long as devices
remain in lower 32 bits. ofmem needs changes for memory ranges and translations.

Note that an enlarged #size-cells is not strictly needed
and constitutes a difference between Apple and IBM firmware (1 vs. 2).

v2:
* Introduce local helper push_physaddr(). A more general, global definition
  was not easily reachable due to inability to determine in preprocessor
  whether phys_addr_t is wider than ucell, leading to compilation errors
  due to a shift the size of the variable.

Signed-off-by: Andreas Färber <andreas.faerber@web.de>


git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@975 f158a5a8-5612-0410-a976-696ce0be7e32