ppc: avoid runtime relocations
Because the sizes of cells and pointers do not match on PPC64,
relocations of an ucell array using the address of the array itself
can't be computed by the linker.

Use a fixed address on PPC64 for the array, so we can use a fixed
ucell sized constant for the relocation offset.

Signed-off-by: Blue Swirl <blauwirbel@gmail.com>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1018 f158a5a8-5612-0410-a976-696ce0be7e32