Fix level 14 and level 10 timers under Solaris 8 on SPARC32.
During boot, if the level 14 timer limit is not set then the Solaris kernel panics with a division by zero
error. Fix this by setting it to the default value as documented in obio.h. Also if the level 10 timer is
not set with a limit of 0, the Solaris kernel constantly outputs "spurious interrupt at processor level
10" messages during boot.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@siriusit.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1033 f158a5a8-5612-0410-a976-696ce0be7e32