ppc: fix mtmsr for SF setting
When running openbios-ppc32 on a ppc64 VM, we need to unset SF on every
interrupt to ensure that things still work.

So far we've been using mtmsr for this operation, but according to the
spec mtmsr doesn't set any bit above 32bit, so we were merely exploiting
a bug in qemu before.

This patch adds a runtime check on MSR_SF to see if we're running on a 64-bit
capable CPU. If so, we use mtmsrd, which can set the high 32bits of MSR.

CC: Andreas Färber <andreas.faerber@web.de>
CC: Segher Boessenkool <segher@kernel.crashing.org>
Signed-off-by: Alexander Graf <agraf@suse.de>

---

v1 -> v2:

  - use segher's awesome MSR_SF detection

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1043 f158a5a8-5612-0410-a976-696ce0be7e32