ppc: only clear MSR_SF when clearing high MSR bits
We are pretty aggressive now with how we clear the high MSR bits, clearing
all of them when we find a 64-bit CPU.

As Segher points out, this is not always a good idea though. There might
be bits set that can be crucial for operation. Hence we should only clear
MSR_SF, as that's the bit we're interested in.

Reported-by: Segher Boessenkool <segher@kernel.crashing.org>
Signed-off-by: Alexander Graf <agraf@suse.de>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1044 f158a5a8-5612-0410-a976-696ce0be7e32