SPARC64: Remove majority of the video initialisation hack.
Now that we have OFMEM, as long as we have a PCI bus then we can configure
the address of the framebuffer automatically without having an
architecture-specific hack. The only thing we need to do is override the
virtual address from the OFMEM phys == virt default.

Similarly we can remove all reference to the video memory variables _vmem
and _evmem since the memory is allocated outside of the OpenBIOS image.

This commit also fixes another couple of issues: switch video.c to use virtual
instead of physical addresses for architectures that require it (the default
is the existing behaviour where phys_addr == virt_addr) and also correct the
framebuffer size calculation which was also adding the framebuffer start address
to size of the framebuffer to be mapped.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1054 f158a5a8-5612-0410-a976-696ce0be7e32