Add a default "decode-unit" word for devices that don't implement their own.
When a device does not implement its own "encode-unit" word, OpenBIOS
currently supplies a default implementation that encodes a hex string
to a single cell.

This commit does the same for "decode-unit" ensuring that it is also
possible to open a device path containing a unit address generated by
a device that doesn't implement its own "encode-unit" word.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1074 f158a5a8-5612-0410-a976-696ce0be7e32