PPC: Fix mapping of OpenBIOS ROM in RAM copy within OFMEM
The OpenBIOS ROM in RAM copy was being incorrectly mapped in OFMEM as part of
the 1:1 physical to virtual mapping used for OpenBIOS housekeeping at the top
of RAM.

Change the existing 1:1 mapping to handle everything up to the ROM in RAM copy,
and then add a separate virtual mapping for the OpenBIOS ROM in RAM address
space.

This ensures that the OpenBIOS address space mapping is now recorded in the MMU
translations property, and hence presented to client kernels to be incorporated
in their existing mapping list before taking control of the MMU.

With this patch applied, my NetBSD 5 test ISO is now able to complete a
substantial part of its kernel boot.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1082 f158a5a8-5612-0410-a976-696ce0be7e32