adb_kbd.c: Implement dummy get-key-map word for the ADB keyboard package.
This word is used by some bootloaders to detect keypresses at startup
e.g. to enable verbose boot. The current implementation simply returns
an empty array (indicating no keys are pressed) as the key/bitmap
mapping is unknown; however this allows bootloaders which make use of
the word to execute.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1083 f158a5a8-5612-0410-a976-696ce0be7e32