mac-parts.c: Update bootpath to reflect the chosen partition if unspecified.
When not booting from a specified partition, bootpath doesn't contain the
selected partition id. Since this is parsed by BootX in order to locate the
Mach kernel, update it accordingly. Otherwise BootX falls back to its default
behaviour of scanning the first 3 partitions in order to find the kernel and
fails on early Darwin/OS X images.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1089 f158a5a8-5612-0410-a976-696ce0be7e32