Remove /chosen "mmu" property from the default device tree (second attempt).
Not all architectures have an "mmu" property in /chosen, and having a NULL
value causes the internal property formatting nodes to invoke a NULL ihandle
reference. The architectures that require it already add it into the device
tree as required during initialisation, so there should be no change in
behaviour. As reported by Artyom Tarasenko.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1117 f158a5a8-5612-0410-a976-696ce0be7e32