SPARC32: Remove high memory virtual memory reservation
The high memory virtual memory reservation introduced in r1114 was a mistake.
We need to leave this area as available for clients but simply not allocate
from it so that the 0xffc00000 area is free for kadb on older versions of
SunOS. Since this part is already handled by the introduction of OFMEM_VIRT_TOP
in the same commit then the behaviour is already correct. As reported by Artyom
Tarasenko.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1130 f158a5a8-5612-0410-a976-696ce0be7e32