video: switch VGA and TCX drivers over to Forth console using is-install.
Here we install the Forth terminal console instead of the C console for
the VGA and TCX drivers.

Note that this patch also disables video output for SPARC32 when
CONFIG_DEBUG_CONSOLE_VIDEO is enabled, since these routines write directly
to the screen instead of using the Forth bindings causing some display
issues.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1156 f158a5a8-5612-0410-a976-696ce0be7e32