PPC: Fix quik bootloader on New World Macs
Currently quik assumes a fixed load-base of 0x3f4000, and only worked on Old
World Macs because of a special exception in bootcode_load.c. Since quik seems
to be the only user of %BOOT known to date, widen the check to all Apple
machines as per conversation with Alex. PReP machines will continue to use the
standard value of load-base as normal.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1165 f158a5a8-5612-0410-a976-696ce0be7e32