load: fix generation of /chosen bootpath property
It seems that some OS require the bootpath property to be the full, unaliased
path to the selected boot device, so use get-instance-path to ensure that
this is the case.

There is a slight issue here when auto-probing a partition, since the package
arguments won't contain the chosen partition id which semms to be mandatory
for bootpath. In order to resolve this, scan for a special
selected-partition-args property in the instance chain before
generating the bootpath, and if it exists use that for the device package
arguments instead.

Using this mechanism, we can set this property once we have auto-probed a
partition in mac-parts.c so bootpath is generated correctly. Finally it is
possible to remove the current hack of overwriting bootpath when calling
open-dev on a mac-parts partition, which would confuse some bootloaders
if they read bootpath after opening a new device.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1167 f158a5a8-5612-0410-a976-696ce0be7e32