video_common.c: split video setup routines into setup_video()
Leave per-device work in init_video() for the moment.

As part of this commit we fix up switch-arch so that CONFIG_QEMU works for
SPARC32 - if we are given the framebuffer information via the QEMU firmware
interface we use it. Otherwise fall back to the values of the VGA_DEFAULT_*
constants.

This commit requires a corresponding patch for QEMU as currently the
framebuffer width/height are not passed via the firmware interface for
SPARC32.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1182 f158a5a8-5612-0410-a976-696ce0be7e32