terminal.fs: fix handling of backspace (0x8) control code
Annex B.2 of the IEEE1275 specification mentions that this code moves the
cursor one position to the left on the current line (i.e. it shouldn't
attempt to redraw any characters).

Remove the character redraw code to bring in line with the specification
which also has the side-effect of fixing the spinner used by quik.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1198 f158a5a8-5612-0410-a976-696ce0be7e32