display: move creation of "display" and "screen" properties to Forth
This means we can remove this code from vga_vbe_init() and move it
in to the generic display system. As part of this, we alter the
"input" and "output" words so that they update the stdin/stdout
handles under /chosen.

While we are here, fix up the various architectures so that they
don't call "input" and "output" themselves - this is now handled
by the Forth console.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1210 f158a5a8-5612-0410-a976-696ce0be7e32