sbus/tcx: implement probe_self() wrapper
Implement a C wrapper around the Forth probe-self word and test it by moving
TCX package properties to tcx.fs.

Note that the probe-self-sbus word is currently hard-coded to invoke the TCX
driver FCode - this is only temporary and will be replaced when we have a
working cpeek implementation.

Also remove the creation of the SUNW,tcx package from tree.fs since this is
now done as part of probe-self-sbus.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1212 f158a5a8-5612-0410-a976-696ce0be7e32