PCI: switch calculation of host PCI memory base to use an offset
Instead of using two different variables to track the absolute values of the
PCI memory base and the host memory base, change the host memory base to be
an offset which is added to the PCI memory base during conversion. As a
consequence of this, the code in pci_bus_addr_to_host_addr() can be unified
to be consistent across all architectures.

Note that this commit also renames arch->host_mem_base to arch->host_pci_base
in order to better describe its function.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1226 f158a5a8-5612-0410-a976-696ce0be7e32