forthstrap: avoid unaligned accesses
Trying to compile OpenBIOS for SPARC32 on a SPARC machine, I get a
SIGBUS within forthstrap, more precisely accessing reloc_table in
load_dictionary. This table is not aligned because the dictionnary
head as a size which is not a multiple of 4.

This is due to the file QEMU,tcx.bin which is being encoded, and which
has a size which is not a multiple of 4. Fix that by adding some padding
after the encoded data.

Signed-off-by: Aurelien Jarno <aurelien@aurel32.net>
Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1236 f158a5a8-5612-0410-a976-696ce0be7e32