SPARC32: fix NULL pointer reference when booting from device with no arguments
When booting from a device with no arguments (such as "disk"), the resulting
partition id string would be NULL causing an exception when setting up the
romvec slice identifier. In the case where no slice is specified, we simply
set the partition to 0 in order to preserve existing behaviour.

This fixes booting from hard disk on QEMU as reported by Paul Wilhelm and
Aurelien Jarno.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1238 f158a5a8-5612-0410-a976-696ce0be7e32