sun-parts.c: only interpose ufs-file-system if an argument (filename) is specified
The code to interpose ufs-file-system similar to OBP would always interpose the
package if it existed, causing Solaris ufsboot to fail when the package was
subsequently opened on a raw disk device without a filename. Fix the code so that
interposition only occurs when a filename is passed as an argument, which enables
us to complete the load and start executing the Solaris 9 kernel on SPARC64.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1245 f158a5a8-5612-0410-a976-696ce0be7e32