SPARC32: copy the a.out header beneath load-base during load
NextSTEP's bootloader for Sparc32 expects that the 32 bytes a.out header
is copied just before the start address.

Since this is a compatibility hack for an implementation detail of OBP then
implement the functionality within a #define AOUT_HEADER_COPY section so
it can easily be removed if people really need a strictly compliant firmware.

Signed-off-by: Olivier Danet <odanet@caramail.com>
Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1265 f158a5a8-5612-0410-a976-696ce0be7e32