OFMEM: ensure minimum alignment of PAGE_SIZE
NextStep appears to request an alignment of 4 for some of its large memory
ranges, and since this is an exact power of 2 then the code to ensure a
minimum alignment of PAGE_SIZE is accidentally bypassed. Move this to a
separate check so we can guarantee the minimum alignment requirement is always
observed.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1268 f158a5a8-5612-0410-a976-696ce0be7e32