OFMEM: redefine OFMEM_TRACE macro so that it always compiles
Make sure that we always compile the OFMEM_TRACE macro regardless of whether
CONFIG_DEBUG_OFMEM is set, similar to the way in which the DEBUG macros work
in QEMU. This relies on the fact that the compiler optimiser will work out
that the compile-time "if" test is always true/false and optimise the result
out if is it not required.

This ensures that we always get proper debug argument checking and means that
certain sections of code don't keep giving unused variable warnings when they
only appear in debug codepaths.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1285 f158a5a8-5612-0410-a976-696ce0be7e32