SPARC64: preserve window state when invoking CIF functions
Members of the *BSD family have a restriction that no window spill/fill traps
can occur in the short time between calling SUNW,set-trap-table in OF and
returning back to the caller.

In order to ensure we minimise window spill/fill traps, preserve the entire
window state across the CIF calls (probably similar to OBP) to ensure that
we don't breach this restriction.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://coreboot.org/openbios/trunk/openbios-devel@1302 f158a5a8-5612-0410-a976-696ce0be7e32