interpreter.fs: don't clobber stack across evaluate strings split on newlines
When an evaluate string is split across a newline, the current string
position is assumed to be the top-most item on the stack. However
in the case of yaboot, items are left on the stack to be used by a
subsequent line within the same evaluate statement and so subsequent
lines are parsed incorrectly.

Fix this by saving the current string position on the r-stack across calls to
(evaluate) so the stack remains correct for subsequent lines.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>