mac: Don't set AAPL,address properties on NewWorld
They are created by the MacOS ROM and in some cases MacOS X to contain
the *virtual* addresses of the BARs as mapped by the kernel.

Having these already there with the wrong values causes the hangs
with MacOS X 10.5 and the video driver.

The reason is that the IONDRVFramebuffer code in OS X 10.5 won't replace
those properties if they already exist, so the driver ends up accessing
the addresses put in there by OpenBIOS which aren't valid mapped virtual
addresses under OS X.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>