hfsplus: don't force OS 9 startup folder selection
For some reason the HFS+ code will ignore the startup folder chosen by the user
and force the OS 9 startup folder if it exists. This can break OS X installations
with Classic, making it impossible to boot back into the OS.

Removing the code to force the use of the OS 9 startup folder allows switching
the startup disk between OS 9/OS X using the Disk Startup Utility in each OS
as expected.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>