mac-parts.c: restrict checking of partition status flags to CHRP partitions only
Early versions of MacOS X don't correctly set the partition status flags when
creating the OS partition. Relax the status flag checks so they are only applied
to CHRP partitions which allows us to both meet the intention of the CHRP
specification and also enable OpenBIOS to boot OS partitions from OSs such as
MacOS X DPs.

With thanks to Steven Troughton-Smith <steve@highcaffeinecontent.com> for the
in-depth investigation.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>