pci: fix setup of 64-bit BARs
The current PCI implementation is designed for a 32-bit PCI bus. When
a 64-bit BAR is encountered during PCI bus enumeration, currently we
treat the MSB BAR the same as a standard 32-bit BAR, attempting to guess
its size by setting to 0xffffffff and reading back the response which
returns back an invalid size of 0 and fails.

This has the effect of setting the MSB BAR to 0xffffffff meaning that
64-bit BARs are mapped at the top of PCI memory in the wrong location.

Handle this case by ignoring the MSB BAR when we detect a 64-bit BAR
leaving it at its default value of zero. This solves the issue by
ensuring that all 64-bit BARs are mapped within 32-bit PCI space which
is fine for the PPC/SPARC64 machines we care about.

Most noticeably this solves the issue of virtio modern devices (which
use 64-bit BARs under QEMU) causing a kernel panic upon startup on
qemu-system-sparc64 because they are mapped at the wrong address by OpenBIOS.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>