vga.fs: install MacOS VGA PPC driver from QEMU fw_cfg interface if available
If the fw_cfg interface provides a MacOS VGA PPC NDRV binary driver then
install it into the "driver,AAPL,MacOS,PowerPC" property of the QEMU VGA
package.

The installation of the binary is configured via the NVRAM configuration
variable "vga-ndrv?" (which is set to true by default for PPC) and also the
presence of the NDRV binary driver file within the QEMU fw_cfg interface
itself.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>