pci: don't add simple open/close words to all PCI devices
Currently all PCI devices except display devices are initialised with
simple open and close words that are effectively a no-op, but enable
the device to be accessed.

Unfortunately this means that if devices have custom open and close words
then the existing ones are overwritten causing OpenBIOS to emit warnings
such as "open isn't unique" on the console.

Resolve this by defaulting all PCI devices to the empty PCI node template
ob_pci_empty_node and then adding the simple open and close words after
the device has been initialised if they do not already exist.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>