display.fs: fix palette entries for foreground and background colors
Some versions of Solaris assume that the foreground color is stored in palette
index 255 and the background color is stored in index 0. Make sure we do the
same otherwise the Solaris video driver picks the background color for console
text which ends up being invisible.

Note that this already matches the default palettes in cgthree.fs and tcx.fs
which were obtained from Sun documentation.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>